food-menu-api-flask/requirements.txt
Flask
pymongo[srv]
Flask-pymongo
python-dotenv